ui/requirements.txt
pillow
requests
streamlit